refactor: create generic `SelectableCard`
- This replaces `TrackCardWrapper` and decouple it from `TrackCard`.
- Fix not playing when `TrackHistoryCard` popup menu play button is pressed
